requirements.txt
arrow==1.2.3
binaryornot==0.4.4
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
defusedxml==0.7.1
django-allauth==0.53.1
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
MarkupSafe==2.1.2
networkx==3.0
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
plotly==5.13.1
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.2
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
tenacity==8.2.2
text-unidecode==1.3
textblob==0.17.1
textblob-fr==0.2.0
tqdm==4.65.0
tweepy==4.13.0
urllib3==1.26.15
wincertstore==0.2
